Use a web worker
