+++ src/utils/useGetChangelog.js
@@ -1,28 +1,22 @@
 import { useEffect, useState } from 'react';
+// eslint-disable-next-line import/no-webpack-loader-syntax
+import * as createWorker from 'workerize-loader!../parse';
+
+let worker = createWorker();
+
+let defaultMarkdown = { type: 'all', ast: { type: 'root', children: [] } };
 
 export function useGetChangelog(filePath, noChangelogFilename) {
-  const [changelog, updateChangelog] = useState('');
-  const [isLoading, setLoading] = useState(false);
-  const [errorMessage, setErrorMessage] = useState();
+  const [state, setState] = useState({
+    changelog: defaultMarkdown,
+    isLoading: true,
+  });
 
   useEffect(() => {
     if (filePath && !noChangelogFilename) {
-      fetch(filePath)
-        .then(res => res.text())
-        .then(text => {
-          setLoading(false);
-          if (text.status === 'error') {
-            setErrorMessage(text);
-          } else {
-            updateChangelog(text);
-          }
-        })
-        .catch(err => {
-          setLoading(false);
-          setErrorMessage(err);
-        });
+      worker.getMarkdown(filePath).then(x => setState(x));
     }
   }, [filePath, noChangelogFilename]);
 
-  return { changelog, isLoading, errorMessage };
+  return state;
 }
